Refactored ZIP to exist independant of brickftp
