+++ zip/zip_test.go
@@ -1,19 +1,23 @@
-package brickftp_test
+package zip_test
 
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/HomesNZ/go-common/brickftp"
 	"path"
+	"github.com/HomesNZ/go-common/zip"
+	"os"
 )
 
 var _ = Describe("Download", func() {
 	Context("zip archive", func() {
 		It("extracts a zip file", func() {
-			val, err := brickftp.UnpackZIP("test/Archive.zip")
+			val, err := zip.UnpackZIP("test/Archive.zip")
 			Expect(err).NotTo(HaveOccurred())
 
 			Expect(path.Base(val)).To(Equal("Archive"))
+			
+			err = os.RemoveAll(val)
+			Expect(err).NotTo(HaveOccurred())
 		})
 	})
 })
